moved snapData into federal folder
