+++ src/programs/state/massachusetts/snap.js
@@ -5,7 +5,7 @@ import {
   getGrossUnearnedIncomeMonthly
 } from '../../../utils/cashflow';
 import { Result } from '../../../utils/Result';
-import { data } from '../../../data/state/massachusetts/undated/snap';
+import { data } from '../../../data/federal/snap';
 import { getYearlyLimitBySize, getMonthlyLimitBySize } from '../../../utils/getGovData';
 import { federalPovertyGuidelines } from '../../../data/federal/federalPovertyGuidelines';
